+++ lib/create_climb/create_climb_event.dart
@@ -1,0 +1,12 @@
+abstract class CreateClimbEvent {}
+
+class ClearGradesEvent extends CreateClimbEvent {}
+
+class GradesUpdatedEvent extends CreateClimbEvent {
+  GradesUpdatedEvent(this.grades);
+
+  final List<String> grades;
+}
+
+class CreateClimbErrorEvent extends CreateClimbEvent {}
+