+++ lib/location/location_view.dart
@@ -5,30 +5,30 @@ import 'package:sendrax/models/climb.dart';
 import 'package:sendrax/models/location.dart';
 import 'package:sendrax/navigation_helper.dart';
 import 'package:sendrax/util/constants.dart';
+import 'package:uuid/uuid.dart';
 
 import 'location_bloc.dart';
 import 'location_state.dart';
 
 class LocationScreen extends StatefulWidget {
-  LocationScreen({Key key, @required this.displayName, @required this.locationId})
+  LocationScreen({Key key, @required this.location})
       : super(key: key);
 
-  final String displayName;
-  final String locationId;
+  final SelectedLocation location;
 
   @override
-  State<StatefulWidget> createState() => _LocationState(locationId);
+  State<StatefulWidget> createState() => _LocationState(location);
 }
 
 class _LocationState extends State<LocationScreen> {
-  final String locationId;
+  final SelectedLocation location;
 
-  _LocationState(this.locationId);
+  _LocationState(this.location);
 
   @override
   Widget build(BuildContext context) {
     return BlocProvider<LocationBloc>(
-      create: (context) => LocationBloc(locationId),
+      create: (context) => LocationBloc(location),
       child: LocationWidget(
         widget: widget,
         widgetState: this,
@@ -48,11 +48,11 @@ class LocationWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.displayName),
+        title: Text(widget.location.displayName),
         actions: <Widget>[
           IconButton(
             icon: Icon(Icons.edit),
-            onPressed: () => _editLocation(widget.locationId, widget.displayName),
+            onPressed: () => _editLocation(widget.location.id, widget.location.displayName, widget.location.gradesId),
           )
         ],
       ),
@@ -82,25 +82,8 @@ class LocationWidget extends StatelessWidget {
                 itemCount: state.sections.length,
               );
             }
-            return _wrapContentWithFab(context, content);
+            return _wrapContentWithFab(state, context, content);
           }),
-    );
-  }
-
-  Widget _wrapContentWithFab(BuildContext context, Widget content) {
-    return Stack(
-      children: <Widget>[
-        content,
-        Container(
-          alignment: Alignment.bottomRight,
-          padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
-          child: FloatingActionButton(
-              onPressed: null,
-              child: Icon(Icons.add, color: Colors.white),
-              backgroundColor: Colors.pinkAccent,
-              elevation: UIConstants.STANDARD_ELEVATION),
-        )
-      ],
     );
   }
 
@@ -135,12 +118,39 @@ class LocationWidget extends StatelessWidget {
     }
   }
 
+  Widget _wrapContentWithFab(LocationState state, BuildContext context, Widget content) {
+    return Stack(
+      children: <Widget>[
+        content,
+        Container(
+          alignment: Alignment.bottomRight,
+          padding: EdgeInsets.all(UIConstants.STANDARD_PADDING),
+          child: FloatingActionButton(
+              onPressed: () => _createClimb(state),
+              child: Icon(Icons.add, color: Colors.white),
+              backgroundColor: Colors.pinkAccent,
+              elevation: UIConstants.STANDARD_ELEVATION),
+        )
+      ],
+    );
+  }
+
   ClimbItem _buildItem(Climb climb) {
     return ClimbItem(climb: climb);
   }
 
-  void _editLocation(String locationId, String displayName) {
-    Location location = new Location(locationId, displayName);
+  void _createClimb(LocationState state) {
+    var uuid = new Uuid();
+    // the null values for grade and section here are required as they are used as the initial
+    // values for the dropdowns
+    Climb climb = new Climb("climb-${uuid.v1()}", "", state.locationId, null, state.gradesId, null, false,
+        <String>[]);
+    NavigationHelper.navigateToCreateClimb(widgetState.context, climb, state.sections, false,
+        addToBackStack: true);
+  }
+
+  void _editLocation(String locationId, String displayName, String gradesId) {
+    Location location = new Location(locationId, displayName, gradesId);
     NavigationHelper.navigateToCreateLocation(widgetState.context, location, true,
         addToBackStack: true);
   }