+++ lib/navigation_helper.dart
@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 
 import 'climb/climb_view.dart';
+import 'create_climb/create_climb_view.dart';
 import 'create_location/create_location_view.dart';
 import 'location/location_view.dart';
 import 'login/login_view.dart';
 import 'main/main_view.dart';
+import 'models/climb.dart';
 import 'models/location.dart';
 
 class NavigationHelper {
@@ -24,29 +26,53 @@ class NavigationHelper {
     }
   }
 
-  static void navigateToLocation(BuildContext context, String displayName, String locationId,
+  static void navigateToLocation(BuildContext context, SelectedLocation location,
       {bool addToBackStack: false}) {
     if (addToBackStack) {
       Navigator.push(
           context,
           MaterialPageRoute(
               builder: (context) =>
-                  LocationScreen(displayName: displayName, locationId: locationId)));
+                  LocationScreen(location: location)));
     } else {
       Navigator.pushReplacement(
           context,
           MaterialPageRoute(
               builder: (context) =>
-                  LocationScreen(displayName: displayName, locationId: locationId)));
+                  LocationScreen(location: location)));
     }
   }
 
-  static void navigateToCreateLocation(BuildContext context, Location location, bool isEdit, {bool addToBackStack: false}) {
+  static void navigateToCreateLocation(BuildContext context, Location location, bool isEdit,
+      {bool addToBackStack: false}) {
     if (addToBackStack) {
-      Navigator.push(context, MaterialPageRoute(builder: (context) => CreateLocationScreen(location: location, isEdit: isEdit)));
+      Navigator.push(
+          context,
+          MaterialPageRoute(
+              builder: (context) => CreateLocationScreen(location: location, isEdit: isEdit)));
     } else {
       Navigator.pushReplacement(
-          context, MaterialPageRoute(builder: (context) => CreateLocationScreen(location: location, isEdit: isEdit)));
+          context,
+          MaterialPageRoute(
+              builder: (context) => CreateLocationScreen(location: location, isEdit: isEdit)));
+    }
+  }
+
+  static void navigateToCreateClimb(
+      BuildContext context, Climb climb, List<String> sections, bool isEdit,
+      {bool addToBackStack: false}) {
+    if (addToBackStack) {
+      Navigator.push(
+          context,
+          MaterialPageRoute(
+              builder: (context) =>
+                  CreateClimbScreen(climb: climb, availableSections: sections, isEdit: isEdit)));
+    } else {
+      Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(
+              builder: (context) =>
+                  CreateClimbScreen(climb: climb, availableSections: sections, isEdit: isEdit)));
     }
   }